mmc: support eMMC driver strength configuration
eMMC v4.5 specifies a method to change the card driver strengths to
better match board impedance requirements.  The strength types roughly
match those in the SD v3.00 spec, but are configured differently
and do not take advantage of the SDHCI implementation.  In addition,
eMMC v5.0 introduced a new driver strength type that has no equivalent
in the current SD spec.  The two functionally map as follows:

  SD   eMMC   Impedance   Relative Drive
  B    0      50 ohms     x1  (default)
  A    1      33 ohms     x1.5
  C    2      66 ohms     x0.75
  D    3      100 ohms    x0.5
  x    4      40 ohms     x1.2

This commit implements card drive strengths for eMMC.  Because the
four SD driver strength types are closely coupled to the SDHCI
HOST_CONTROL2 register and it has no extra, adjacent bits available
for the new drive strengths defined by eMMC v5.0, a new group of
driver types are defined.  This commit does take advantage of the
existing CAP bits for SD driver types A-D.

Change-Id: Ic7f9555431b7e0b4784e39b55a256bbec85facb4
Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Reviewed-by: Jeffrey Carlyle <jeff.carlyle@motorola.com>
Reviewed-by: Igor Kovalenko <igork@motorola.com>

# Conflicts:
#	drivers/mmc/core/mmc.c

Signed-off-by: Gobinda Joy <gobinda.joy@gmail.com>