mmc: don't blindly retry on timed-out flushes
The commit "mmc: block: flush request requeue after timeout"
introduced logic that blindly requeues flush requests that have
timed-out without checking the card state or taking advantage of
any of the error recovery logic in the driver.  This leads to an
infinite loop if the card continues to time out because of how
flush requests are prioritized over other requests.

This commit disables the retry except in the case of Hynix cards.

Signed-off-by: Russ Knize <Russ.Knize@motorola.com>
Reviewed-by: Igor Kovalenko <igork@motorola.com>

# Conflicts:
#	drivers/mmc/card/block.c

Signed-off-by: Gobinda Joy <gobinda.joy@gmail.com>