FS: big revert of all iov_iter mods from 3.16.y so called optimizations.
It's not really seems to do any change, but it's pain in $#@ to update code
from 3.10.y when there is big mess with this iov_iter all over the FS and MM!

Dump all of! from all changed files based on history of this kernel.

Now i can update in peace and fix some more stuff.

I have seen reports that show mem failed in ii_iovec_copy_to_user
and in file_read_iter_actor when accesing fuse and other files.

# Conflicts:
#	fs/Makefile
#	fs/exfat/exfat_super.c
#	fs/ext4/indirect.c
#	include/linux/fs.h
#	mm/filemap.c

Signed-off-by: Gobinda Joy <gobinda.joy@gmail.com>