cpu_power: Avoids race condition when the task exits.
When the task is terminated, the cpu_power for that particular task is
added to the terminated tasks. cpu_power is set to ULLONG_MAX at this
point to avoid double accounting of the power. It is possible that
before the task releases all the resources, cpu reschedules the task or
a timer interrupt is fired. At this point we will try to add the
additional time to the process, which will cause the accounting to be
skewed. This avoids the case where we change the cpu_power when it is
already set to ULLONG_MAX.

Bug: 22064385
Change-Id: I405733725d535b0a864088516bf52fa3638ee6aa
Signed-off-by: Jin Qian <jinqian@google.com>
Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>