battery: max17048: Use iadc to get the CURRENT_NOW & CURRENT_AVG
Initial commit to get the battery CURRENT_NOW & CURRENT_AVG value using
iadc.

As we all know max17048 ic lacks any proper way to determine battery
current information. Here comes iadc to the rescue. However, take this
value with a bag of salt as this only provides an idea of CURRENT
information.

Will love an experts opinion of the codes.

[Gobinda: adapted to hlte from Hammerhead]
Signed-off-by: Gobinda Joy <gobinda.joy@gmail.com>