ext4: Move work from io_end to inode
It does not make much sense to have struct work in ext4_io_end_t because
we always use it for only one ext4_io_end_t per inode (the first one in
the i_completed_io list). So just move the structure to inode itself.
This also allows for a small simplification in processing io_end
structures.

Signed-off-by: Jan Kara <jack@xxxxxxx>
Signed-off-by: Gobinda Joy <gobinda.joy@gmail.com>