ext4: Simplify list handling in ext4_do_flush_completed_IO()
The function splices i_completed_io_list to its private list first. From
that moment on we don't need any lock for working with io_end structures
because all io_end structure on the list are only our own. So we can
remove the other two lists in the function and free io_end immediately
after we are done with it.

CC: Dmitry Monakhov <dmonakhov@xxxxxxxxxx>
Signed-off-by: Jan Kara <jack@xxxxxxx>
Signed-off-by: Gobinda Joy <gobinda.joy@gmail.com>