ext4: Fix ext4_writepage() to achieve data=ordered guarantees
So far ext4_writepage() skipped writing pages that had any delayed or
unwritten buffers attached. When blocksize < pagesize this breaks
data=ordered mode guarantees as we can have a page with one freshly
allocated buffer whose allocation is part of the committing transaction
and another buffer in the page which is delayed or unwritten. So fix
this problem by calling ext4_bio_writepage() anyway. It will submit
mapped buffers and leave others alone.

Signed-off-by: Jan Kara <jack@xxxxxxx>
Signed-off-by: Gobinda Joy <gobinda.joy@gmail.com>